Solve the problem that `LookupFileId` lookup urls is empty due to leader switching
The vidMap structure is modified to a linked list structure (the length is limited to 5). When the vidMap is reset, the current vidMap is added to the new vidMap as a cache node. When the query locations is empty, the cache node is searched to avoid problems when the master switches leaders.

(cherry picked from commit 58c02d6429530d674d4e0e493e9e1a860208017e)